Initial commit for CLI branch - balance working transfers started
General flow for the CLI is working but the transfer function still needs some debugging. Made some changes to the contracts, including removing structures as it looked like the other versions of those files were more recent/integral. Attention should be given to ensure that nothing important has been removed before this branch would be considered for an upstream merge.